requirements.txt
uvicorn[standard]==0.34.3
asyncpg==0.30.0
SQLAlchemy==2.0.41
pydantic==2.11.5
pydantic-settings==2.9.1
fastapi==0.115.12
python-baseconv==1.2.2
pyjwt==2.10.1
python-multipart==0.0.20
pwdlib[argon2]==0.2.1

#migrations
alembic==1.16.1

#tests
# httpx==0.28.1
# pytest==8.3.5
# pytest-postgresql==7.0.2
# pytest-subtests==0.14.1
# psycopg==3.2.9
# faker==37.3.0

